backend/requirements.txt
flask
flask-cors
werkzeug
python-dotenv
pytesseract
pdf2image
Pillow
opencv-python-headless
psycopg2-binary
groq
gunicorn
pdfminer.six